requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
tenacity>=8.3.0
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.9.0
seaborn>=0.13.2
tqdm>=4.66.4
ipykernel>=6.29.5
jupyter>=1.0.0
notebook>=7.2.1
pyyaml>=6.0.2